Refactor in preparation to add toObservable
